+++ examples/src/components/ChartView.tsx
@@ -35,9 +35,8 @@ export function ChartView({ className }: ChartViewProps) {
 
       mainIndicatorsRef.current.forEach((name) => {
         chart.createIndicator(
-          { name, id: `main_${name}`, paneId: "candle_pane" },
-          true,
-          { id: "candle_pane" },
+          { name, id: `main_${name}` },
+          { isStack: true, pane: { id: "candle_pane" } },
         );
       });